feat: add lazy loaded page content, remove page item therefore
