--- system/PageItem.php
@@ -1,52 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace herbie;
-
-use ArrayAccess;
-
-final class PageItem implements ArrayAccess
-{
-    use PageItemTrait;
-
-    public function __construct(array $data = [])
-    {
-        $this->initData($data);
-    }
-
-    /**
-     * @param mixed $offset
-     */
-    public function offsetExists($offset): bool
-    {
-        return $this->__isset($offset);
-    }
-
-    /**
-     * @param mixed $offset
-     * @return mixed
-     */
-    #[\ReturnTypeWillChange]
-    public function offsetGet($offset)
-    {
-        return $this->__get($offset);
-    }
-
-    /**
-     * @param mixed $offset
-     * @param mixed $value
-     */
-    public function offsetSet($offset, $value): void
-    {
-        $this->__set($offset, $value);
-    }
-
-    /**
-     * @param mixed $offset
-     */
-    public function offsetUnset($offset): void
-    {
-        throw new \BadMethodCallException('Unset is not supported.');
-    }
-}
