+++ website/site/themes/default/css/styles.css
@@ -265,9 +265,6 @@ table {
     /*color:#00CA4C;*/
     color:#bb2222;
 }
-.sidebar ul ul {
-    margin-bottom:1em;
-}
 .sidebar .menu li a {
     font-weight:bold;
 }
